functions/raw-parse-scoreboard/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-secret-manager
duckdb==1.3.2
pandas==2.3.2